chore: extract prerequisite mod to file
